Main Module: Delete MainModel, change the logic of work with model
