--- SurfSummerSchoolProject/MainFlow/MainModule/Model/MainModel.swift
@@ -1,51 +1,0 @@
-//
-//  MainModel.swift
-//  SurfSummerSchoolProject
-//
-//  Created by Zhasur Sidamatov on 04/08/22.
-//
-
-import Foundation
-import UIKit
-
-final class MainModel {
-    //MARK: - Events
-    var didItemsUpdated: (() -> Void)?
-    
-    //MARK: - Properties
-    let pictureService = PicturesService()
-    var errorDescription: String?
-    static let shared = MainModel()
-    var items: [ItemModel] = []
-    
-    //MARK: - Initialization
-    
-    private init() {}
-    
-    //MARK: - Methods
-    
-    func getPictures(_ completinHandler: @escaping () -> Void) {
-        pictureService.loadPictures { result in
-            switch result {
-            case .success(let pictures):
-                DispatchQueue.main.async {
-                    self.items = pictures.map({ picture in
-                        return ItemModel(imageUrlString: picture.photoUrl,
-                                         title: picture.title,
-                                         isFavorite: false, //TODO - Create Service
-                                         dataCreation: "15.05.2022",
-                                         description: picture.content)
-                    })
-                    completinHandler()
-                }
-                
-            case .failure(let error):
-                self.errorDescription = error.localizedDescription
-                DispatchQueue.main.async {
-                    completinHandler()
-                }
-            }
-        }
-    }
-    
-}
