+++ SurfSummerSchoolProject/MainFlow/MainModule/Presenter/MainPresenter.swift
@@ -8,19 +8,27 @@
 import Foundation
 
 
-class MainPresenter: MainViewOutput {
+class MainPresenter {
+    
+    //MARK: - Properties
     
     weak var view: MainViewInput?
     var router: MainRouterInput?
-    var model = MainModel.shared
+    let pictureService: PicturesService = .init()
+    var items: [ItemModel] = []
+    var errorDescription: String?
     
-    
+}
+
+//MARK: - MainViewOutput
+
+extension MainPresenter: MainViewOutput {
     func showDetailViewController(item: ItemModel) {
         router?.showDetailModule(item: item)
     }
     
     func loadPosts(_ completion: @escaping () -> Void) {
-        model.getPictures(completion)
+        getPictures(completion)
     }
     
     func showSearchViewController() {
@@ -32,4 +40,27 @@ class MainPresenter: MainViewOutput {
 
 private extension MainPresenter {
 
+    func getPictures(_ completinHandler: @escaping () -> Void) {
+        pictureService.loadPictures { result in
+            switch result {
+            case .success(let pictures):
+                DispatchQueue.main.async {
+                    self.items = pictures.map({ picture in
+                        return ItemModel(imageUrlString: picture.photoUrl,
+                                         title: picture.title,
+                                         isFavorite: false, //TODO - Create Service
+                                         dataCreation: "15.05.2022",
+                                         description: picture.content)
+                    })
+                    completinHandler()
+                }
+                
+            case .failure(let error):
+                self.errorDescription = error.localizedDescription
+                DispatchQueue.main.async {
+                    completinHandler()
+                }
+            }
+        }
+    }
 }
